Add common/fr.* files
