+++ src/common/common.c
@@ -264,81 +264,6 @@ C_KZG_RET blob_to_polynomial(fr_t *p, const Blob *blob) {
     return C_KZG_OK;
 }
 
-////////////////////////////////////////////////////////////////////////////////////////////////////
-// Field Operations
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Test whether two field elements are equal.
- *
- * @param[in]   a   The first element
- * @param[in]   b   The second element
- *
- * @retval true     The two elements are equal.
- * @retval false    The two elements are not equal.
- */
-bool fr_equal(const fr_t *a, const fr_t *b) {
-    uint64_t _a[4], _b[4];
-    blst_uint64_from_fr(_a, a);
-    blst_uint64_from_fr(_b, b);
-    return _a[0] == _b[0] && _a[1] == _b[1] && _a[2] == _b[2] && _a[3] == _b[3];
-}
-
-/**
- * Test whether the operand is one in the finite field.
- *
- * @param[in] p The field element to be checked
- *
- * @retval true  The element is one
- * @retval false The element is not one
- */
-bool fr_is_one(const fr_t *p) {
-    uint64_t a[4];
-    blst_uint64_from_fr(a, p);
-    return a[0] == 1 && a[1] == 0 && a[2] == 0 && a[3] == 0;
-}
-
-/**
- * Divide a field element by another.
- *
- * @param[out] out `a` divided by `b` in the field
- * @param[in]  a   The dividend
- * @param[in]  b   The divisor
- *
- * @remark The behavior for `b == 0` is unspecified.
- * @remark This function supports in-place computation.
- */
-void fr_div(fr_t *out, const fr_t *a, const fr_t *b) {
-    blst_fr tmp;
-    blst_fr_eucl_inverse(&tmp, b);
-    blst_fr_mul(out, a, &tmp);
-}
-
-/**
- * Exponentiation of a field element.
- *
- * Uses square and multiply for log(n) performance.
- *
- * @param[out] out `a` raised to the power of `n`
- * @param[in]  a   The field element to be exponentiated
- * @param[in]  n   The exponent
- *
- * @remark A 64-bit exponent is sufficient for our needs here.
- * @remark This function does support in-place computation.
- */
-void fr_pow(fr_t *out, const fr_t *a, uint64_t n) {
-    fr_t tmp = *a;
-    *out = FR_ONE;
-
-    while (true) {
-        if (n & 1) {
-            blst_fr_mul(out, out, &tmp);
-        }
-        if ((n >>= 1) == 0) break;
-        blst_fr_sqr(&tmp, &tmp);
-    }
-}
-
 /**
  * Compute and return [ x^0, x^1, ..., x^{n-1} ].
  *
