+++ src/eip7594/fft.c
@@ -15,6 +15,7 @@
  */
 
 #include "fft.h"
+#include "common/fr.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 // FFT Functions for Field Elements
